Enhancements

+++ app/src/main/java/com/zaed/reservationmanager/ui/company/details/components/BalanceSection.kt
@@ -103,7 +103,7 @@ fun BalanceSection(
 fun IncomeExpenseCardSection(
     modifier: Modifier = Modifier,
     title: String,
-    amount: Double
+    amount: Int
 ) {
     Column(
         modifier = modifier,
