+++ app/src/main/java/com/zaed/reservationmanager/ui/util/StringUtil.kt
@@ -4,7 +4,7 @@ import java.text.NumberFormat
 import java.util.Currency
 import java.util.Locale
 
-fun Double.formatMoney(padding: Int = 0): String {
+fun Int.formatMoney(padding: Int = 0): String {
     val formatter = NumberFormat.getCurrencyInstance(Locale.getDefault()).apply {
         minimumFractionDigits = padding
         maximumFractionDigits = padding
